requirements.txt
playwright
pandas
mysql-connector-python
openpyxl
python-dotenv
